refactor: refactor Accordion component and remove Divider component
