+++ src/components/data/Accordion/styles.ts
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 import { inube } from "@inube/design-system";
+import { IDividerProps } from ".";
 
 interface IStyledContainer {
   isMobile: boolean;
@@ -30,4 +31,15 @@ export const StyledHead = styled.button`
   align-items: center;
   width: 100%;
   cursor: pointer;
+`;
+
+export const StyledDivider = styled.hr<IDividerProps>`
+  margin: 0;
+  width: 100%;
+  height: 0px;
+  border: none;
+  border-top: 0.5px ${({ dashed }) => (dashed ? "dashed" : "solid")};
+  border-top-color: ${({ theme }) =>
+    theme.color?.stroke?.divider?.regular ||
+    inube.color.stroke.divider.regular};
 `;
